Make it possible to develop using Docker
Add Dockerfiles and docker-compose.yml that sets up two
containers, one for Node and one for the MySQL database.

After installing Docker and Docker Compose you should be able to
run "docker-compose up" to build and run the containers, and
start the development server.

You will likely need to run "docker-compose up db" first. During
the first run downloaded data is imported into the database which
takes a while. The Node container doesn't yet wait for this
initialization to finish and fails.